moved test source for class_ref up one dir
