+++ app/src/test/resources/class_ref/ExternalClass.java
@@ -1,0 +1,4 @@
+package class_ref;
+public class ExternalClass {
+    public int CONSTANT = 0;
+}
