--- app/src/test/resources/class_ref/src/ExternalClass.java
@@ -1,3 +1,0 @@
-public class ExternalClass {
-    
-}
