--- app/src/test/resources/class_ref/src/MyClass.java
@@ -1,13 +1,0 @@
-import java.lang.Exception;
-import com.ExternalClass;
-
-public class MyClass {
-    
-	public void method() {
-		try {
-			this.call(ExternalClass.class);
-		} catch (Exception t) {
-			throw new Exception("ignore", t);
-		}
-	}
-}
